requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.0

# Database & Cache
# supabase==2.3.0 
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Environment & Configuration
python-dotenv==1.0.1
pydantic-settings

# Utilities
httpx==0.27.0
websockets==12.0
python-slugify==8.0.4
loguru==0.7.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.5
requests==2.31.0